feat: move the last scan cache into VoucherScanner instead. Only 1 check function is then passed into VoucherScanner
